+++ sources/sort.c
@@ -2957,7 +2957,7 @@ NoPoly:
 
 /*
  		#] Compare1 : 
- 		#[ CompareSymbols :			int CompareSymbols(term1,term2,par)
+ 		#[ CompareSymbols :			WORD CompareSymbols(term1,term2,par)
 */
 /**
  *	Compares the terms, based on the value of AN.polysortflag.
@@ -2972,7 +2972,7 @@ NoPoly:
  *	to that of Compare1 which is the regular compare routine in sort.c
  */
 
-int CompareSymbols(WORD *term1, WORD *term2, WORD par)
+WORD CompareSymbols(WORD *term1, WORD *term2, WORD par)
 {
 	GETIDENTITY
 	int sum1, sum2;
@@ -3005,7 +3005,7 @@ int CompareSymbols(WORD *term1, WORD *term2, WORD par)
 
 /*
  		#] CompareSymbols : 
- 		#[ CompareHSymbols :		int CompareHSymbols(term1,term2,par)
+ 		#[ CompareHSymbols :		WORD CompareHSymbols(term1,term2,par)
 */
 /**
  *	Compares terms that can have only SYMBOL and HAAKJE subterms.
@@ -3016,7 +3016,7 @@ int CompareSymbols(WORD *term1, WORD *term2, WORD par)
  *	to that of Compare1 which is the regular compare routine in sort.c
  */
 
-int CompareHSymbols(WORD *term1, WORD *term2, WORD par)
+WORD CompareHSymbols(WORD *term1, WORD *term2, WORD par)
 {
 	GETIDENTITY
 	WORD *t1, *t2, *tt1, *tt2, *ttt1, *ttt2;
